requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
click==8.1.7
h11==0.14.0
idna==3.10
joblib==1.4.2
mysql-connector-python==9.1.0
nltk==3.9.1
outcome==1.3.0.post0
PySocks==1.7.1
regex==2024.11.6
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
zeyrek==0.1.3